Cleaning, Visulaizing and trying out diffrent classifiers on the data
Customer Service is one of top priorities for Aigües de Barcelona. We
are in the process of redesigning our customer services to improve our
relationship with the customers. To make sure, we respond to our
customers efficiently, we need to predict the Volume and typologies of
contacts to the call center. Further, we also need to forecast the
number of resolutions opened on a daily basis. Your task is to forecast
the number of contacts (by various contact types) and number of
resolutions (by category and Subject) opened on a daily basis for Aigües
de Barcelona.

Contacts: A contact is defined whenever a customer initiates and gets in
touch with the Company. The contact can happen through various channels
including phone, mail, visit, fax etc.
Resolutions: A Resolution is created every time a customer has placed a
request with the company. This could be through a customer contact (as
defined above) or otherwise (for example, when the company called the
customer